prod

+++ main_app/views.py
@@ -1,14 +1,39 @@
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import render, get_object_or_404, redirect
 from .forms import *
 from .models import *
 from django.http import HttpResponseRedirect
 from django.contrib import messages
 from itertools import chain
 from django.db.models import Q
+from datetime import date
+from main_app.management.commands.utils import get_qiwi_balance
+from django.urls import reverse_lazy
+from django.contrib.auth.views import LoginView
+from django.contrib.auth import logout
 
+
+class LoginUser(LoginView):
+    """В приложении не используется, можно использовать для стандартной аутентификации"""
+    form_class = LoginUserForm
+    template_name = 'main_app/login.html'
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['title'] = 'Логин'
+        return context
+
+    def get_success_url(self):
+        return reverse_lazy('all_product')
+
+
+def logout_user(request):
+    logout(request)
+    return redirect('login')
 
 def index(request):
     '''Вывод всех товаров на главной странице'''
+    if not request.user.is_authenticated:
+        return redirect('login')
     product = Product.objects.all()
     category = Category.objects.all()
     if 'search' in request.GET:
@@ -28,7 +53,6 @@ def index(request):
             product = product.filter(count__gte=int(params['from_count']))
         if 'to_count' in params:
             product = product.filter(count__lte=int(params['to_count']))
-
         if 'order_by' in params:
             product = product.order_by(params['order_by'])
     return render(request, 'main_app/index.html', {'product': product, 'category': category})
@@ -36,20 +60,22 @@ def index(request):
 
 def product_view(request, pk):
     '''Страница изменения продукта/добавления кол-ва в товар и списание'''
+    if not request.user.is_authenticated:
+        return redirect('login')
     product = get_object_or_404(Product, pk=pk)
     category = Category.objects.all()
     if request.method == "POST" and 'update' in request.POST:
         product_form = Product_reqForm(request.POST, files=request.FILES, instance=product)
         if product_form.is_valid():
             product_form.save()
-            return HttpResponseRedirect('/')
+            return redirect('productdetail', pk=pk)
 
     if request.method == "POST" and 'delete' in request.POST:
         messages.info(request, 'Вы уверены?')
 
     if request.method == 'GET' and 'confirm_delete' in request.GET:
         product.delete()
-        return HttpResponseRedirect('/')
+        return redirect('all_product')
 
     if request.method == 'POST':
         if 'reception' in request.POST:
@@ -61,19 +87,25 @@ def product_view(request, pk):
                 f = form.save(commit=False)
                 f.product = product
                 f.save()
-                return HttpResponseRedirect('/')
+                return redirect('productdetail', pk=pk)
 
         elif 'liquidated' in request.POST:
             count = int(request.POST['liquidated_count'])
             note = request.POST['liquidated_note']
-            ReceptionProduct.objects.create(price=0, count=count, note=note, product=product, liquidated=True)
+            ReceptionProduct.objects.create(price=int(request.POST['liquidated_price']), count=count, note=note, product=product, liquidated=True)
             product.count -= count
             product.save()
-            return HttpResponseRedirect('/')
+            return redirect('productdetail', pk=pk)
 
     reception_form = ReceptionForm()
     reception_queryset = ReceptionProduct.objects.filter(product__pk=pk)
     sold_queryset = SoldProduct.objects.filter(product__pk=pk)
+
+    sold_stat = (sum([x.count * x.price for x in sold_queryset]), sum([x.count for x in sold_queryset]))
+    reception_stat = (sum([x.count * x.price for x in reception_queryset]), sum([x.count for x in reception_queryset]))
+    liquidated_stat = sum([x.count * x.price for x in reception_queryset.filter(liquidated=True)]) ,sum([x.count for x in reception_queryset.filter(liquidated=True)])
+    all_stat = sold_stat[0] - reception_stat[0] - liquidated_stat[0]
+    
 
     params = {k:v for k,v in request.GET.items() if len(v) != 0}
 
@@ -97,31 +129,45 @@ def product_view(request, pk):
                                         'weight': product.weight,
                                         'subcategory': product.subcategory_id})
                                         
-    return render(request, 'main_app/product.html', {'product_form': product_form, 'product': product, 'category': category, 'reception_form': reception_form, 'trade_queryset': result_list})
+    return render(request, 'main_app/product.html', {'product_form': product_form,
+                                                     'product': product,
+                                                     'category': category,
+                                                     'reception_form': reception_form,
+                                                     'trade_queryset': result_list,
+                                                     'sold_stat': sold_stat,
+                                                     'reception_stat': reception_stat, 
+                                                     'liquidated_stat': liquidated_stat,
+                                                     'all_stat': all_stat,
+                                                     })
 
 
 def create_product(request):
     '''Страница создания товара'''
+    if not request.user.is_authenticated:
+        return redirect('login')
     if request.method == "POST":
         product_form = ProductForm(request.POST, files=request.FILES)
         if product_form.is_valid():
             product = product_form.save()
             product.slug = f'p||{product.pk}'
             product.save()
-            return HttpResponseRedirect('/')
+            return redirect('productdetail', pk=product.pk)
 
     product_form = ProductForm(initial={'count': 0})
     category = Category.objects.all()
     return render(request, 'main_app/add_product.html', {'product_form': product_form, 'category': category})
 
 
-def reception_product(request, pk):
-    product = get_object_or_404(Product, pk=pk)
+def reception_product(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
     reception_form = ReceptionForm()
     return render(request, 'main_app/reception.html', {'reception_form': reception_form})
 
 
 def create_category(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
     if request.method == 'POST' and 'create_category' in request.POST:
         category_form = CategoryForm(request.POST, files=request.FILES)
         if category_form.is_valid():
@@ -129,10 +175,10 @@ def create_category(request):
             model.slug = f'c||{model.pk}'
             model.save()
             messages.info(request, 'Новая категория успешно создана!')
-            return HttpResponseRedirect('/add_category/')
+            return redirect('add_category')
         else:
             messages.info(request, 'Ошибка!')
-            return HttpResponseRedirect('/add_category/')
+            return redirect('add_category')
 
     elif request.method == "POST" and 'create_sc' in request.POST:
         sc_form = SubcategoryForm(request.POST, files=request.FILES)
@@ -143,7 +189,7 @@ def create_category(request):
             f.slug = f'sc||{f.pk}'
             f.save()
             messages.info(request, 'Новая категория успешно создана!')
-            return HttpResponseRedirect('/add_category/')
+            return redirect('add_category')
 
     category_form = CategoryForm()
     sc_form = SubcategoryForm()
@@ -153,12 +199,14 @@ def create_category(request):
 
 
 def category_view(request, pk):
+    if not request.user.is_authenticated:
+        return redirect('login')
     category = get_object_or_404(Category, pk=pk)
     if request.method == "POST":
         category_form = Category_reqForm(request.POST, files=request.FILES, instance=category)
         if category_form.is_valid():
             category_form.save()
-        return HttpResponseRedirect('/add_category')
+        return redirect('add_category')
     
     if request.method == 'GET' and 'delete' in request.GET:
         messages.info(request, 'Вы уверены?')
@@ -166,7 +214,7 @@ def category_view(request, pk):
     elif request.method == 'GET' and 'confirm_delete' in request.GET:
         messages.info(request, 'Категория успешно удалена')
         category.delete()
-        return HttpResponseRedirect('/add_category/')
+        return redirect('add_category')
 
 
     category_form = Category_reqForm(initial={'name': category.name, 'max_count_product': category.max_count_product})
@@ -174,12 +222,14 @@ def category_view(request, pk):
 
 
 def subcategory_view(request, pk):
+    if not request.user.is_authenticated:
+        return redirect('login')
     subcategory = get_object_or_404(SubCategory, id=pk)
     if request.method == "POST":
         category_form = Subcategory_reqForm(request.POST, files=request.FILES, instance=subcategory)
         if category_form.is_valid():
             category_form.save()
-        return HttpResponseRedirect('/add_category')
+        return redirect('add_category')
     
     if request.method == 'GET' and 'delete' in request.GET:
         messages.info(request, 'Вы уверены?')
@@ -187,13 +237,15 @@ def subcategory_view(request, pk):
     elif request.method == 'GET' and 'confirm_delete' in request.GET:
         messages.info(request, 'Категория успешно удалена')
         subcategory.delete()
-        return HttpResponseRedirect('/add_category/')
+        return redirect('add_category')
 
     category_form = Subcategory_reqForm(initial={'name': subcategory.name})
     return render(request, 'main_app/category_detail.html', {'category': subcategory, 'category_form': category_form})
 
 
 def new_order(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
     if request.method == 'POST':
         order_id = int(request.POST['order_id'])
         try:
@@ -205,10 +257,13 @@ def new_order(request):
         order.track_code = track_code
         order.check_admin = True
         order.save()
+    title = 'Новые заказы'
     queryset = OrderingProduct.objects.filter(check_admin=False)
-    return render(request, 'main_app/order.html', context={'queryset': queryset})
+    return render(request, 'main_app/order.html', context={'queryset': queryset, 'title': title})
 
 def old_order(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
     if request.method == 'POST':
         order_id = int(request.POST['order_id'])
         try:
@@ -220,22 +275,61 @@ def old_order(request):
         order.track_code = track_code
         order.check_admin = True
         order.save()
-        
+    
+    title = 'Обработанные заказы'
     queryset = OrderingProduct.objects.filter(check_admin=True)
-    return render(request, 'main_app/order.html', context={'queryset': queryset})
+    return render(request, 'main_app/order.html', context={'queryset': queryset, 'title': title})
 
 
 def control_qiwi(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
     if request.method == 'POST' and 'new_token' in request.POST:
         form = QiwiTokenForm(request.POST)
-        if form.is_valid():
-            form.save()
+        try:
+            balance = get_qiwi_balance(request.POST['number'], request.POST['token'])
+            if form.is_valid():
+                qiwi = form.save()
+                qiwi.balance = balance
+                qiwi.save()
+        except:
+            messages.info(request, 'Ошибка получения баланса кошелька')
     elif request.method == 'POST' and 'activate' in request.POST:
         QiwiToken.objects.update(active=False)
         qiwi_id = int(request.POST['radio'])
         QiwiToken.objects.filter(pk=qiwi_id).update(active=True)
+    elif request.method == 'POST' and 'del_tok' in request.POST:
+        qiwi_id = int(request.POST['del_tok_list'])
+        QiwiToken.objects.get(pk=qiwi_id).delete()
 
     pay_product = PayProduct.objects.all()
     queryset = QiwiToken.objects.all()
     form = QiwiTokenForm()
     return render(request, 'main_app/qiwi.html', context={'form': form, 'queryset': queryset, 'pay_product': pay_product})
+
+
+
+def user_stat(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    users = TelegramUser.objects.all()
+    sold_queryset = SoldProduct.objects.all()
+    reception_queryset = ReceptionProduct.objects.all()
+    params = {k:v for k,v in request.GET.items() if v != ''}
+
+    if 'start' in params and 'end' in params:
+        sold_queryset = sold_queryset.filter(date__range=[params['start'], params['end']])
+        reception_queryset = reception_queryset.filter(date__range=[params['start'], params['end']])
+    elif 'start' in params:
+        sold_queryset = sold_queryset.filter(date__gte=params['start'])
+        reception_queryset = reception_queryset.filter(date__gte=params['start'])
+    elif 'end' in params:
+        sold_queryset = sold_queryset.filter(date__lte=params['end'])
+        reception_queryset = reception_queryset.filter(date__lte=params['end'])      
+
+    sold_stat = (sum([x.count * x.price for x in sold_queryset]), sum([x.count for x in sold_queryset]))
+    reception_stat = (sum([x.count * x.price for x in reception_queryset]), sum([x.count for x in reception_queryset]))
+    liquidated_stat = sum([x.count * x.price for x in reception_queryset.filter(liquidated=True)]) ,sum([x.count for x in reception_queryset.filter(liquidated=True)])
+    all_stat = sold_stat[0] - reception_stat[0] - liquidated_stat[0]
+
+    return render(request, 'main_app/user_stat.html', context={'users': users, 'sold_stat': sold_stat, 'reception_stat': reception_stat, 'liquidated_stat': liquidated_stat, 'all_stat': all_stat})
